Connect to VU-AMS only in the first prepare
Several parts of the preparations only have to happen ones (i.e. not for
every vu_ams item in the exp):
1) Loading the dll
2) Connecting to the device
3) Checking if device is recording
4) Appending Close to cleanup_functions
Because the item.py get and set functions can only handle simple
variable types (unicode, float, int);
we can not store the <class 'ctypes.WinDLL'> returned by
windll.amsserial globally.
Therefore we load the dll (1) in every prepare. Luckily it seems to
return the same handle every time.
To track if the device is already connected (so 2, 3 and 4 can be
skipped) we use the global variable
vuamsconnected.